+++ src/main/java/com/gdschongik/gdsc/domain/event/domain/service/EventDomainService.java
@@ -12,7 +12,7 @@ public class EventDomainService {
     /**
      * 행사의 최대 신청자 수를 변경할 때, 현재 신청자 수보다 낮게 설정하는 경우 예외를 발생시킵니다.
      */
-    public void validateWhenUpdateMaxApplicantCount(int currentApplicants, @Nullable Integer newMaxApplicantCount) {
+    public void validateWhenUpdateMaxApplicantCount(long currentApplicants, @Nullable Integer newMaxApplicantCount) {
         if (newMaxApplicantCount == null) { // 인원 제한을 없애는 경우 항상 허용
             return;
         }
